Use a float64 to set the uptime_in_ms gauge metric. (#6532)
A float32 starts to lose precision at 16777216ms which is just under 5
hours.

Fixes https://github.com/spiffe/spire/issues/6048

+++ pkg/common/uptime/uptime.go
@@ -24,7 +24,7 @@ func reportMetrics(ctx context.Context, interval time.Duration, m telemetry.Metr
 	t := clk.Ticker(interval)
 	defer t.Stop()
 	for {
-		telemetry.EmitUptime(m, float32(Uptime()/time.Millisecond))
+		telemetry.EmitUptime(m, float64(Uptime()/time.Millisecond))
 		select {
 		case <-t.C:
 		case <-ctx.Done():
